Create README.md
This is a base project for calculating animals/weather/objects in nature and learning about their behavior. The squirrel moves when the mouse touches it, so it is not bugging out. Future use case is to set these objects in a defined space with a learned behavior to see how the system evolves.